Removed CDW support for Immunization, also I feex Bryan's typo.
