+++ data-query/spotbugs-excludes.xml
@@ -5,11 +5,6 @@
     The transfomer can return an optional boolean matching the objects property per the FHIR specification.
   -->
   <Match>
-    <Class name="gov.va.api.health.dataquery.service.controller.immunization.ImmunizationTransformer"/>
-    <Method name="reported"/>
-    <Bug pattern="NP_BOOLEAN_RETURN_NULL"/>
-  </Match>
-  <Match>
     <Class name="gov.va.api.health.dataquery.service.controller.patient.DatamartPatientTransformer"/>
     <Method name="deceased"/>
     <Bug pattern="NP_BOOLEAN_RETURN_NULL"/>
