+++ data-query/src/main/java/gov/va/api/health/dataquery/service/controller/immunization/Dstu2ImmunizationTransformer.java
@@ -17,7 +17,7 @@ import java.util.Optional;
 import lombok.Builder;
 
 @Builder
-public class DatamartImmunizationTransformer {
+public class Dstu2ImmunizationTransformer {
 
   private final DatamartImmunization datamart;
