--- data-query/src/main/java/gov/va/api/health/dataquery/service/controller/immunization/ImmunizationController.java
@@ -1,327 +1,0 @@
-package gov.va.api.health.dataquery.service.controller.immunization;
-
-import static gov.va.api.health.dataquery.service.controller.Transformers.firstPayloadItem;
-import static gov.va.api.health.dataquery.service.controller.Transformers.hasPayload;
-import static java.util.Collections.emptyList;
-
-import gov.va.api.health.argonaut.api.resources.Immunization;
-import gov.va.api.health.argonaut.api.resources.Immunization.Bundle;
-import gov.va.api.health.dataquery.service.controller.AbstractIncludesIcnMajig;
-import gov.va.api.health.dataquery.service.controller.Bundler;
-import gov.va.api.health.dataquery.service.controller.Bundler.BundleContext;
-import gov.va.api.health.dataquery.service.controller.CountParameter;
-import gov.va.api.health.dataquery.service.controller.PageLinks;
-import gov.va.api.health.dataquery.service.controller.PageLinks.LinkConfig;
-import gov.va.api.health.dataquery.service.controller.Parameters;
-import gov.va.api.health.dataquery.service.controller.ResourceExceptions.NotFound;
-import gov.va.api.health.dataquery.service.controller.Validator;
-import gov.va.api.health.dataquery.service.controller.WitnessProtection;
-import gov.va.api.health.dataquery.service.mranderson.client.MrAndersonClient;
-import gov.va.api.health.dataquery.service.mranderson.client.Query;
-import gov.va.api.health.dstu2.api.resources.OperationOutcome;
-import gov.va.dvp.cdw.xsd.model.CdwImmunization103Root;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.List;
-import java.util.Optional;
-import java.util.function.Function;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-import javax.servlet.http.HttpServletResponse;
-import javax.validation.constraints.Min;
-import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.BooleanUtils;
-import org.apache.commons.lang3.StringUtils;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.util.MultiValueMap;
-import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-
-/**
- * Request Mappings for Immunization Profile, see
- * https://www.fhir.org/guides/argonaut/r2/StructureDefinition-argo-Immunization.html for
- * implementation details.
- */
-@SuppressWarnings("WeakerAccess")
-@Validated
-@RestController
-@RequestMapping(
-  value = {"/dstu2/Immunization"},
-  produces = {"application/json", "application/json+fhir", "application/fhir+json"}
-)
-@Slf4j
-public class ImmunizationController {
-  private final ImmunizationController.Datamart datamart = new ImmunizationController.Datamart();
-  private Transformer transformer;
-  private MrAndersonClient mrAndersonClient;
-  private Bundler bundler;
-  private ImmunizationRepository repository;
-  private WitnessProtection witnessProtection;
-  private boolean defaultToDatamart;
-
-  /** Spring constructor. */
-  @SuppressWarnings("ParameterHidesMemberVariable")
-  public ImmunizationController(
-      @Value("${datamart.immunization}") boolean defaultToDatamart,
-      @Autowired ImmunizationController.Transformer transformer,
-      @Autowired MrAndersonClient mrAndersonClient,
-      @Autowired Bundler bundler,
-      @Autowired ImmunizationRepository repository,
-      @Autowired WitnessProtection witnessProtection) {
-    this.defaultToDatamart = defaultToDatamart;
-    this.transformer = transformer;
-    this.mrAndersonClient = mrAndersonClient;
-    this.bundler = bundler;
-    this.repository = repository;
-    this.witnessProtection = witnessProtection;
-  }
-
-  private Immunization.Bundle bundle(
-      MultiValueMap<String, String> parameters, int page, int count) {
-    CdwImmunization103Root root = search(parameters);
-    LinkConfig linkConfig =
-        LinkConfig.builder()
-            .path("Immunization")
-            .queryParams(parameters)
-            .page(page)
-            .recordsPerPage(count)
-            .totalRecords(root.getRecordCount().intValue())
-            .build();
-    return bundler.bundle(
-        BundleContext.of(
-            linkConfig,
-            root.getImmunizations() == null
-                ? Collections.emptyList()
-                : root.getImmunizations().getImmunization(),
-            transformer,
-            Immunization.Entry::new,
-            Immunization.Bundle::new));
-  }
-
-  /** Read by id. */
-  @GetMapping(value = {"/{publicId}"})
-  public Immunization read(
-      @RequestHeader(value = "Datamart", defaultValue = "") String datamartHeader,
-      @PathVariable("publicId") String publicId) {
-    if (datamart.isDatamartRequest(datamartHeader)) {
-      return datamart.read(publicId);
-    }
-    return transformer.apply(
-        firstPayloadItem(
-            hasPayload(search(Parameters.forIdentity(publicId)).getImmunizations())
-                .getImmunization()));
-  }
-
-  /** Read by id. */
-  @GetMapping(
-    value = {"/{publicId}"},
-    headers = {"raw=true"}
-  )
-  public String readRaw(@PathVariable("publicId") String publicId, HttpServletResponse response) {
-    ImmunizationEntity entity = datamart.readRaw(publicId);
-    AbstractIncludesIcnMajig.addHeader(response, entity.icn());
-    return entity.payload();
-  }
-
-  private CdwImmunization103Root search(MultiValueMap<String, String> params) {
-    Query<CdwImmunization103Root> query =
-        Query.forType(CdwImmunization103Root.class)
-            .profile(Query.Profile.ARGONAUT)
-            .resource("Immunization")
-            .version("1.03")
-            .parameters(params)
-            .build();
-    return hasPayload(mrAndersonClient.search(query));
-  }
-
-  /** Search by _id. */
-  @GetMapping(params = {"_id"})
-  public Immunization.Bundle searchById(
-      @RequestHeader(value = "Datamart", defaultValue = "") String datamartHeader,
-      @RequestParam("_id") String publicId,
-      @RequestParam(value = "page", defaultValue = "1") @Min(1) int page,
-      @CountParameter @Min(0) int count) {
-    if (datamart.isDatamartRequest(datamartHeader)) {
-      return datamart.searchById(publicId, page, count);
-    }
-    return bundle(
-        Parameters.builder()
-            .add("identifier", publicId)
-            .add("page", page)
-            .add("_count", count)
-            .build(),
-        page,
-        count);
-  }
-
-  /** Search by Identifier. */
-  @GetMapping(params = {"identifier"})
-  public Immunization.Bundle searchByIdentifier(
-      @RequestHeader(value = "Datamart", defaultValue = "") String datamartHeader,
-      @RequestParam("identifier") String publicId,
-      @RequestParam(value = "page", defaultValue = "1") @Min(1) int page,
-      @CountParameter @Min(0) int count) {
-    return searchById(datamartHeader, publicId, page, count);
-  }
-
-  /** Search by patient. */
-  @GetMapping(params = {"patient"})
-  public Immunization.Bundle searchByPatient(
-      @RequestHeader(value = "Datamart", defaultValue = "") String datamartHeader,
-      @RequestParam("patient") String patient,
-      @RequestParam(value = "page", defaultValue = "1") @Min(1) int page,
-      @CountParameter @Min(0) int count) {
-    if (datamart.isDatamartRequest(datamartHeader)) {
-      return datamart.searchByPatient(patient, page, count);
-    }
-    return bundle(
-        Parameters.builder().add("patient", patient).add("page", page).add("_count", count).build(),
-        page,
-        count);
-  }
-
-  /** Hey, this is a validate endpoint. It validates. */
-  @PostMapping(
-    value = "/$validate",
-    consumes = {"application/json", "application/json+fhir", "application/fhir+json"}
-  )
-  public OperationOutcome validate(@RequestBody Immunization.Bundle bundle) {
-    return Validator.create().validate(bundle);
-  }
-
-  public interface Transformer
-      extends Function<CdwImmunization103Root.CdwImmunizations.CdwImmunization, Immunization> {}
-
-  /**
-   * This class is being used to help organize the code such that all the datamart logic is
-   * contained together. In the future when Mr. Anderson support is dropped, this class can be
-   * eliminated.
-   */
-  private class Datamart {
-
-    private Bundle bundle(
-        MultiValueMap<String, String> parameters, List<Immunization> reports, int totalRecords) {
-      PageLinks.LinkConfig linkConfig =
-          PageLinks.LinkConfig.builder()
-              .path("Immunization")
-              .queryParams(parameters)
-              .page(Parameters.pageOf(parameters))
-              .recordsPerPage(Parameters.countOf(parameters))
-              .totalRecords(totalRecords)
-              .build();
-      return bundler.bundle(
-          Bundler.BundleContext.of(
-              linkConfig,
-              reports,
-              Function.identity(),
-              Immunization.Entry::new,
-              Immunization.Bundle::new));
-    }
-
-    private Bundle bundle(
-        MultiValueMap<String, String> parameters, int count, Page<ImmunizationEntity> entities) {
-
-      log.info("Search {} found {} results", parameters, entities.getTotalElements());
-      if (count == 0) {
-        return bundle(parameters, emptyList(), (int) entities.getTotalElements());
-      }
-
-      return bundle(
-          parameters,
-          replaceReferences(
-                  entities
-                      .get()
-                      .map(ImmunizationEntity::asDatamartImmunization)
-                      .collect(Collectors.toList()))
-              .stream()
-              .map(this::transform)
-              .collect(Collectors.toList()),
-          (int) entities.getTotalElements());
-    }
-
-    ImmunizationEntity findById(String publicId) {
-      String cdwId = witnessProtection.toCdwId(publicId);
-      Optional<ImmunizationEntity> entity = repository.findById(cdwId);
-      return entity.orElseThrow(() -> new NotFound(publicId));
-    }
-
-    boolean isDatamartRequest(String datamartHeader) {
-      if (StringUtils.isBlank(datamartHeader)) {
-        return defaultToDatamart;
-      }
-      return BooleanUtils.isTrue(BooleanUtils.toBooleanObject(datamartHeader));
-    }
-
-    private PageRequest page(int page, int count) {
-      return PageRequest.of(page - 1, count == 0 ? 1 : count, ImmunizationEntity.naturalOrder());
-    }
-
-    Immunization read(String publicId) {
-      DatamartImmunization immunization = findById(publicId).asDatamartImmunization();
-      replaceReferences(List.of(immunization));
-      return transform(immunization);
-    }
-
-    ImmunizationEntity readRaw(String publicId) {
-      return findById(publicId);
-    }
-
-    Collection<DatamartImmunization> replaceReferences(Collection<DatamartImmunization> resources) {
-      /*
-       * Reaction is not a reference to another resource and is intentionally excluded from
-       * registration.
-       */
-      witnessProtection.registerAndUpdateReferences(
-          resources,
-          resource ->
-              Stream.of(
-                  resource.patient(),
-                  resource.performer().orElse(null),
-                  resource.requester().orElse(null),
-                  resource.encounter().orElse(null),
-                  resource.location().orElse(null)
-                  //
-                  ));
-      return resources;
-    }
-
-    Bundle searchById(String publicId, int page, int count) {
-      Immunization resource = read(publicId);
-      return bundle(
-          Parameters.builder()
-              .add("identifier", publicId)
-              .add("page", page)
-              .add("_count", count)
-              .build(),
-          resource == null || count == 0 ? emptyList() : List.of(resource),
-          resource == null ? 0 : 1);
-    }
-
-    Bundle searchByPatient(String patient, int page, int count) {
-      String icn = witnessProtection.toCdwId(patient);
-      log.info("Looking for {} ({})", patient, icn);
-      return bundle(
-          Parameters.builder()
-              .add("patient", patient)
-              .add("page", page)
-              .add("_count", count)
-              .build(),
-          count,
-          repository.findByIcn(icn, page(page, count)));
-    }
-
-    Immunization transform(DatamartImmunization dm) {
-      return DatamartImmunizationTransformer.builder().datamart(dm).build().toFhir();
-    }
-  }
-}
