--- data-query/src/test/java/gov/va/api/health/dataquery/service/controller/immunization/ImmunizationControllerTest.java
@@ -1,172 +1,0 @@
-package gov.va.api.health.dataquery.service.controller.immunization;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-
-import gov.va.api.health.argonaut.api.resources.Immunization;
-import gov.va.api.health.argonaut.api.resources.Immunization.Bundle;
-import gov.va.api.health.autoconfig.configuration.JacksonConfig;
-import gov.va.api.health.dataquery.service.controller.Bundler;
-import gov.va.api.health.dataquery.service.controller.Bundler.BundleContext;
-import gov.va.api.health.dataquery.service.controller.PageLinks.LinkConfig;
-import gov.va.api.health.dataquery.service.controller.Parameters;
-import gov.va.api.health.dataquery.service.controller.Validator;
-import gov.va.api.health.dataquery.service.mranderson.client.MrAndersonClient;
-import gov.va.api.health.dataquery.service.mranderson.client.Query;
-import gov.va.api.health.dstu2.api.bundle.AbstractBundle.BundleType;
-import gov.va.dvp.cdw.xsd.model.CdwImmunization103Root;
-import gov.va.dvp.cdw.xsd.model.CdwImmunization103Root.CdwImmunizations;
-import gov.va.dvp.cdw.xsd.model.CdwImmunization103Root.CdwImmunizations.CdwImmunization;
-import java.math.BigInteger;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.function.Supplier;
-import javax.validation.ConstraintViolationException;
-import lombok.SneakyThrows;
-import org.junit.Before;
-import org.junit.Test;
-import org.mockito.ArgumentCaptor;
-import org.mockito.Mock;
-import org.mockito.Mockito;
-import org.mockito.MockitoAnnotations;
-import org.springframework.util.MultiValueMap;
-
-@SuppressWarnings("WeakerAccess")
-public class ImmunizationControllerTest {
-  @Mock MrAndersonClient client;
-
-  @Mock ImmunizationController.Transformer tx;
-
-  ImmunizationController controller;
-  @Mock Bundler bundler;
-
-  @Before
-  public void _init() {
-    MockitoAnnotations.initMocks(this);
-    controller = new ImmunizationController(false, tx, client, bundler, null, null);
-  }
-
-  private void assertSearch(Supplier<Bundle> invocation, MultiValueMap<String, String> params) {
-    CdwImmunization103Root root = new CdwImmunization103Root();
-    root.setPageNumber(BigInteger.valueOf(1));
-    root.setRecordsPerPage(BigInteger.valueOf(10));
-    root.setRecordCount(BigInteger.valueOf(3));
-    root.setImmunizations(new CdwImmunizations());
-    CdwImmunization cdwItem1 = new CdwImmunization();
-    CdwImmunization cdwItem2 = new CdwImmunization();
-    CdwImmunization cdwItem3 = new CdwImmunization();
-    root.getImmunizations().getImmunization().addAll(Arrays.asList(cdwItem1, cdwItem2, cdwItem3));
-    Immunization immunization1 = Immunization.builder().build();
-    Immunization immunization2 = Immunization.builder().build();
-    Immunization immunization3 = Immunization.builder().build();
-    when(tx.apply(cdwItem1)).thenReturn(immunization1);
-    when(tx.apply(cdwItem2)).thenReturn(immunization2);
-    when(tx.apply(cdwItem3)).thenReturn(immunization3);
-    when(client.search(Mockito.any())).thenReturn(root);
-
-    Bundle mockBundle = new Bundle();
-    when(bundler.bundle(Mockito.any())).thenReturn(mockBundle);
-
-    Bundle actual = invocation.get();
-
-    assertThat(actual).isSameAs(mockBundle);
-    @SuppressWarnings("unchecked")
-    ArgumentCaptor<BundleContext<CdwImmunization, Immunization, Immunization.Entry, Bundle>>
-        captor = ArgumentCaptor.forClass(BundleContext.class);
-
-    verify(bundler).bundle(captor.capture());
-
-    LinkConfig expectedLinkConfig =
-        LinkConfig.builder()
-            .page(1)
-            .recordsPerPage(10)
-            .totalRecords(3)
-            .path("Immunization")
-            .queryParams(params)
-            .build();
-    assertThat(captor.getValue().linkConfig()).isEqualTo(expectedLinkConfig);
-    assertThat(captor.getValue().xmlItems()).isEqualTo(root.getImmunizations().getImmunization());
-    assertThat(captor.getValue().newBundle().get()).isInstanceOf(Bundle.class);
-    assertThat(captor.getValue().newEntry().get()).isInstanceOf(Immunization.Entry.class);
-    assertThat(captor.getValue().transformer()).isSameAs(tx);
-  }
-
-  private Bundle bundleOf(Immunization resource) {
-    return Bundle.builder()
-        .type(BundleType.searchset)
-        .resourceType("Bundle")
-        .entry(
-            Collections.singletonList(
-                Immunization.Entry.builder()
-                    .fullUrl("http://example.com")
-                    .resource(resource)
-                    .build()))
-        .build();
-  }
-
-  @SuppressWarnings("unchecked")
-  @Test
-  public void read() {
-    CdwImmunization103Root root = new CdwImmunization103Root();
-    root.setImmunizations(new CdwImmunizations());
-    CdwImmunization xmlImmunization = new CdwImmunization();
-    root.getImmunizations().getImmunization().add(xmlImmunization);
-    Immunization item = Immunization.builder().build();
-    when(client.search(Mockito.any())).thenReturn(root);
-    when(tx.apply(xmlImmunization)).thenReturn(item);
-    Immunization actual = controller.read("false", "hello");
-    assertThat(actual).isSameAs(item);
-    ArgumentCaptor<Query<CdwImmunization103Root>> captor = ArgumentCaptor.forClass(Query.class);
-    verify(client).search(captor.capture());
-    assertThat(captor.getValue().parameters()).isEqualTo(Parameters.forIdentity("hello"));
-  }
-
-  @Test
-  public void searchById() {
-    assertSearch(
-        () -> controller.searchById("false", "me", 1, 10),
-        Parameters.builder().add("identifier", "me").add("page", 1).add("_count", 10).build());
-  }
-
-  @Test
-  public void searchByIdentifier() {
-    assertSearch(
-        () -> controller.searchByIdentifier("false", "me", 1, 10),
-        Parameters.builder().add("identifier", "me").add("page", 1).add("_count", 10).build());
-  }
-
-  @Test
-  public void searchByPatient() {
-    assertSearch(
-        () -> controller.searchByPatient("false", "me", 1, 10),
-        Parameters.builder().add("patient", "me").add("page", 1).add("_count", 10).build());
-  }
-
-  @Test
-  @SneakyThrows
-  public void validateAcceptsValidBundle() {
-    Immunization resource =
-        JacksonConfig.createMapper()
-            .readValue(
-                getClass().getResourceAsStream("/cdw/old-immunization-1.03.json"),
-                Immunization.class);
-
-    Bundle bundle = bundleOf(resource);
-    assertThat(controller.validate(bundle)).isEqualTo(Validator.ok());
-  }
-
-  @Test(expected = ConstraintViolationException.class)
-  @SneakyThrows
-  public void validateThrowsExceptionForInvalidBundle() {
-    Immunization resource =
-        JacksonConfig.createMapper()
-            .readValue(
-                getClass().getResourceAsStream("/cdw/old-immunization-1.03.json"),
-                Immunization.class);
-    resource.resourceType(null);
-
-    Bundle bundle = bundleOf(resource);
-    controller.validate(bundle);
-  }
-}
